Drop the unix and win barrel files and move back to {unix,win}.js
Remove the `src/{unix,win}/index.js` (barrel) files in favor of
`src/{unix,win}.js`. This simplifies imports (as barrel files can't be
imported by their parent folder with EcmaScript modules).

Internal refactoring initially lead to the creation of these `index.js`
files [1] because the old `src/{unix/win}.js` needed to be preserved.
Subsequent refactoring removed the `src/{unix/win}.js` files [2], so now
they can be restored once again.

--
1. e183cfda431216c87bae4773f32c8bc527c64f58
2. a0b50d2e0e26598d2db0267c8f6c83da262dad50
